update project likeCount and followCount on server

+++ src/routes/api/unfollowProject.js
@@ -1,16 +1,25 @@
-import { init } from '../../server/mongo.js';
+import { init, response } from '../../server/mongo.js';
 
 export async function post(req, res, next) {
 	const { db } = await init();
 
 	const options = req.body;
 
-	const details = {};
-	details.userId = options.userId;
-	details.projectId = options.projectId;
+	if (options.userId && options.projectId) {
+		let projectUpdateResult = await db.collection('projects').updateOne({ id: options.projectId }, { $inc: { followCount: -1 } });
 
-	const result = await db.collection('follows').deleteMany(details);
+		if (projectUpdateResult) {
+			const details = {};
+			details.userId = options.userId;
+			details.projectId = options.projectId;
 
-	res.writeHead(200, {'Content-Type': 'application/json'});
-	res.end(JSON.stringify(result));
+			const result = await db.collection('follows').deleteMany(details);
+
+			response(res, result);
+		} else {
+			response(res, {error: true});
+		}
+	} else {
+		response(res, {error: true});
+	}
 }
